feature: add verification of authorization in question service

+++ application/src/main/java/cn/edu/nju/service/examService/QuestionServiceImpl.java
@@ -1,13 +1,12 @@
 package cn.edu.nju.service.examService;
 
-import cn.edu.nju.controller.AccountController;
+import cn.edu.nju.dao.courseDAO.IUserCourseDAO;
 import cn.edu.nju.dao.examDAO.IQuestionDAO;
-import cn.edu.nju.dao.userDAO.IUserDAO;
+import cn.edu.nju.info.ResultInfo;
+import cn.edu.nju.info.examInfo.QuestionInfo;
 import cn.edu.nju.model.examModel.QuestionModel;
 import cn.edu.nju.utils.EncryptionUtil;
 import cn.edu.nju.utils.ExcelUtil;
-import cn.edu.nju.info.ResultInfo;
-import cn.edu.nju.info.examInfo.QuestionInfo;
 import jxl.read.biff.BiffException;
 import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,16 +23,23 @@ public class QuestionServiceImpl implements IQuestionService {
 
     private final IQuestionDAO questionDAO;
 
-    private final IUserDAO userDAO;
+    private final IUserCourseDAO userCourseDAO;
 
     @Autowired
-    public QuestionServiceImpl(IQuestionDAO questionDAO, IUserDAO userDAO) {
+    public QuestionServiceImpl(IQuestionDAO questionDAO,
+                               IUserCourseDAO userCourseDAO) {
         this.questionDAO = questionDAO;
-        this.userDAO = userDAO;
+        this.userCourseDAO = userCourseDAO;
     }
 
     @Override
-    public ResultInfo saveQuestion(int courseId, InputStream excelStream) {
+    public ResultInfo saveQuestion(int userId, int courseId, InputStream excelStream) {
+        if (!userCourseDAO.doesUserHaveCourse(userId, courseId)) {
+            return new ResultInfo(
+                    false, "只有该门课的老师才能上传试题", null
+            );
+        }
+
         try {
             ByteArrayOutputStream bufferStream = toByteArrayOutputStream(excelStream);
             byte[] data = bufferStream.toByteArray();
@@ -64,7 +70,16 @@ public class QuestionServiceImpl implements IQuestionService {
 
     @Override
     @SuppressWarnings("unchecked")
-    public ResultInfo getAllQuestions(int num) {
+    public ResultInfo getAllQuestions(Integer page, Integer size) {
+        if (page == null || page <= 0) {
+            page = 1;
+        }
+        if (size == null || size <= 0) {
+            size = 10;
+        }
+        int num = (long)page * size > Integer.MAX_VALUE ?
+                Integer.MAX_VALUE : page * size;
+
         ResultInfo result = questionDAO.getAllQuestions(num);
         if (!result.isSuccess()) {
             return result;
@@ -83,7 +98,19 @@ public class QuestionServiceImpl implements IQuestionService {
     }
 
     @Override
-    public ResultInfo deleteQuestions(List<Integer> questionIdList) {
+    public ResultInfo deleteQuestions(int userId, List<Integer> questionIdList) {
+        if (questionIdList.isEmpty()) {
+            return new ResultInfo(true, "成功删除试题", null);
+        }
+
+        int questionId = questionIdList.get(0);
+        int courseId = questionDAO.getCourseIdByQuestionId(questionId);
+        if (!userCourseDAO.doesUserHaveCourse(userId, courseId)) {
+            return new ResultInfo(
+                    false, "只有该门课的老师才能删除试题", null
+            );
+        }
+
         return questionDAO.deleteQuestions(questionIdList);
     }
